pyproject.toml
[tool.poetry]
name = "ooerbot"
version = "0.0.0"
description = "Random things for the r/Ooer Discord"
authors = ["William McKinnerney <me@williammck.net>"]
license = "MIT"

[tool.poetry.dependencies]
aiofile = "^3.5.1"
"discord.py" = "^1.7.3"
python = "^3.9"
PyYAML = "^5.4.1"

[tool.poetry.dev-dependencies]
bandit = "^1.7.0"
flake8 = "^3.9.2"
flake8-import-order = "^0.18.1"
flake8-bugbear = "^21.4.3"
flake8-builtins = "^1.5.3"
flake8-quotes = "^3.2.0"
mccabe = "^0.6.1"
mypy = "^0.910"
pep8-naming = "^0.12.0"
pre-commit = "^2.13.0"
types-PyYAML = "^5.4.3"

[tool.poetry.scripts]
bot = "run:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
